+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/Baker.java
@@ -2,6 +2,7 @@ package com.concordium.sdk.responses.accountinfo;
 
 import com.concordium.sdk.crypto.ed25519.ED25519PublicKey;
 import com.concordium.sdk.responses.AccountIndex;
+import com.concordium.sdk.responses.BakerId;
 import com.concordium.sdk.transactions.CCDAmount;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
@@ -23,7 +24,7 @@ public final class Baker {
      * The baker id.
      */
     @JsonProperty("bakerId")
-    private final AccountIndex bakerId;
+    private final BakerId bakerId;
 
     /**
      * The staked amount.