+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/poolstatus/PendingChangeRemovePool.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.extern.jackson.Jacksonized;
 
+import java.time.OffsetDateTime;
 import java.util.Date;
 
 @Data
@@ -14,5 +15,5 @@ public class PendingChangeRemovePool extends PendingChange {
     /**
      * Effective time of the change.
      */
-    private final Date effectiveTime;
+    private final OffsetDateTime effectiveTime;
 }