+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Added support for GRPC V2 `GetPassiveDelegators` for getting the passive delegators at a certain block.
 - Added support for GRPC V2 `GetPoolDelegators` for getting the delegators for a specified baking pool at a certain block.
 - Added support for GRPC V2 `GetElectionInfo` for getting the bakers at a certain block and other metadata of the election.
+- Added support for GRPC V2 `GetPoolInfo` for gettting information about a given pool at the end of a given block.
 
 ## 4.2.0
 - Added initial support for GRPC V2